+++ src/Products/PloneMeeting/migrations/migrate_to_4103.py
@@ -7,8 +7,11 @@ from Products.PloneMeeting.migrations import Migrator
 
 class Migrate_To_4103(Migrator):
 
-    def _fixAdministratorsGroup(self):
-        """Make sure the 'Administrators' group exists and has the 'Manager' role."""
+    def _fixGroupsAndUsersRoles(self):
+        """Make sure the 'Administrators' group exists and has the 'Manager' role.
+           Make sure every source_users have the 'Member' role and
+           every source_groups does not have the 'Member' role,
+           including the 'AuthenticatedUsers' group."""
         logger.info("Fixing \"Administrators\" group to have the \"Manager\" role...")
         # create 'Administrators' if it does not exist
         group_name = 'Administrators'
@@ -24,17 +27,33 @@ class Migrate_To_4103(Migrator):
                             "this was fixed!".format(group_name))
             else:
                 logger.info("\"{0}\" group had already \"Manager\" role.".format(group_name))
+
+        # manage users, make sure we do not get LDAP users, listMembers to not get LDAP users
+        # give 'Member' role to every users
+        logger.info("Fixing users, make sure every has the 'Member' role...")
+        membership = api.portal.get_tool('portal_membership')
+        members = membership.listMembers()
+        for member in members:
+            if 'Member' not in member.getRoles():
+                api.user.grant_roles(user=member, roles=['Member'])
+
+        # remove 'Member' role from every groups, including 'AuthenticatedUsers'
+        logger.info("Fixing groups, make sure no more group has the 'Member' role...")
+        group_ids = [grp.id for grp in api.group.get_groups()]
+        assert('AuthenticatedUsers' in group_ids)
+        for group_id in group_ids:
+            api.group.revoke_roles(groupname=group_id, roles=['Member'])
         logger.info('Done.')
 
     def run(self):
         logger.info('Migrating to PloneMeeting 4103...')
-        self._fixAdministratorsGroup()
+        self._fixGroupsAndUsersRoles()
 
 
 def migrate(context):
     '''This migration function will:
 
-       1) Fix 'Administrators' group to have the 'Manager' role.
+       1) Fix groups and users roles (roles 'Manager' and 'Member').
     '''
     migrator = Migrate_To_4103(context)
     migrator.run()